Works without ImageMagick
If ImageMagick is not installed, the script will use a pre-converted PNG
image (if it exists).
Even if ImageMagick is installed, the script can work with the PNG
images instead of SVG images.
The script will auto detect which image file is installed.
If the PNG file does not exist and ImageMagick is not installed, then it
will just use the SVG image.